Moving repositories and model to api
