+++ api/models/Series.js
