+++ setup/setupAdmin.js
@@ -1,4 +1,5 @@
-const User = require('../web/models/User')
+// TODO: Remove this. It should call the repository
+const User = require('../api/models/User')
 
 module.exports = (app) => {
   if (process.env.NODE_ENV === 'DEV') {
@@ -11,6 +12,7 @@ module.exports = (app) => {
   } else {
     const productionAdminEmail = process.env.ADMIN_EMAIL
     const productionAdminPass = process.env.ADMIN_PASS
+
     return createAdminUser({
       email: productionAdminEmail,
       password: productionAdminPass
@@ -35,16 +37,19 @@ function createAdminUser (data) {
     throw new Error('Admin user can not have null values')
   }
 
+  // TODO: Remove this. It should call the repository
   User.findOne({
     email: email
   }).then((user) => {
     if (user) {
       return;
     }
+
     const newAdminUser = new User({
       email: email,
       password: passd
     })
+
     newAdminUser.save(err => {
       if (err) {
         console.log('Can\'t create default user')
